Add get balance x loan id

+++ src/main/java/com/mercadocredito/domain/payment/IPaymentRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IPaymentRepository extends JpaRepository<Payment,Integer> {
+public interface IPaymentRepository extends JpaRepository<Payment, Long> {
 }
